+++ lib/liquex/filter.ex
@@ -31,12 +31,12 @@ defmodule Liquex.Filter do
       ) do
     func = String.to_existing_atom(function)
 
-    function_args =
-      Enum.map(
-        arguments,
-        &Liquex.Argument.eval(&1, context)
-      )
-      |> merge_keywords()
+    {function_args, context} =
+      Enum.map_reduce(arguments, context, fn arg, ctx ->
+        Liquex.Argument.eval(arg, ctx)
+      end)
+
+    function_args = merge_keywords(function_args)
 
     mod =
       if mod != __MODULE__ and Kernel.function_exported?(mod, func, length(function_args) + 2) do
@@ -46,7 +46,8 @@ defmodule Liquex.Filter do
       end
 
     # Liquid's filters operate on Ruby Arrays; ranges go through `to_a` first.
-    Kernel.apply(mod, func, [normalize(value) | function_args] ++ [context])
+    result = Kernel.apply(mod, func, [normalize(value) | function_args] ++ [context])
+    {result, context}
   rescue
     # credo:disable-for-next-line
     ArgumentError -> raise Liquex.Error, "Invalid filter #{function}"
@@ -119,8 +120,12 @@ defmodule Liquex.Filter do
 
   defp do_at_least(v, m) do
     cond do
-      Math.nan?(v) -> m
-      Math.nan?(m) -> v
+      Math.nan?(v) ->
+        m
+
+      Math.nan?(m) ->
+        v
+
       Math.special?(v) or Math.special?(m) ->
         if Math.compare(v, m) == :gt, do: v, else: m
 
@@ -151,8 +156,12 @@ defmodule Liquex.Filter do
 
   defp do_at_most(v, m) do
     cond do
-      Math.nan?(v) -> m
-      Math.nan?(m) -> v
+      Math.nan?(v) ->
+        m
+
+      Math.nan?(m) ->
+        v
+
       Math.special?(v) or Math.special?(m) ->
         if Math.compare(v, m) == :lt, do: v, else: m
 
@@ -1094,12 +1103,7 @@ defmodule Liquex.Filter do
 
   def sum(list, property, _) when is_list(list) do
     Enum.reduce(list, 0, fn item, acc ->
-      value =
-        cond do
-          is_map(item) -> Liquex.Indifferent.get(item, property, 0)
-          true -> 0
-        end
-
+      value = if is_map(item), do: Liquex.Indifferent.get(item, property, 0), else: 0
       Math.add(acc, to_number(value))
     end)
   end