+++ lib/liquex/tag/for_tag.ex
@@ -248,11 +248,10 @@ defmodule Liquex.Tag.ForTag do
       ) do
     {resolved_offset, parameters} = resolve_offset(parameters, collection, context)
 
+    {value, context} = Liquex.Argument.eval(collection, context)
+
     items =
-      collection
-      |> Liquex.Argument.eval(context)
-      |> Expression.eval_collection(parameters)
-      |> case do
+      case Expression.eval_collection(value, parameters) do
         nil -> nil
         c -> Collection.to_enumerable(c) |> Enum.to_list()
       end
@@ -341,15 +340,6 @@ defmodule Liquex.Tag.ForTag do
     do: %{context | private: Map.put(private, continue_key(collection), value)}
 
   defp forloop(index, length, parentloop) do
-    %{
-      "index" => index + 1,
-      "index0" => index,
-      "rindex" => length - index,
-      "rindex0" => length - index - 1,
-      "first" => index == 0,
-      "last" => index == length - 1,
-      "length" => length,
-      "parentloop" => parentloop
-    }
+    %Liquex.Drop.Forloop{index: index, length: length, parentloop: parentloop}
   end
 end